# apps/web/app/api/revisions/[n]/[...path]/route.ts
import { NextResponse, type NextRequest } from 'next/server';
import { createApiContext } from '@/server/api/session';
import { pathParamSchema, versionParamSchema, parseParams, getPathFromParams, formatZodError } from '@/server/api/validate';
import { apiError, mapDomainError, internalError } from '@/server/api/errors';
import { DomainError } from '@/server/errors';
import { withApiAudit, type RouteHandler } from '@/server/api/audit-wrapper';
import * as pageService from '@/server/services/pages';

/**
 * Get a specific revision.
 *
 * @openapi
 * @summary Get a revision
 * @description Returns a specific revision of a page by path and version number.
 * @tag Revisions
 * @response RevisionView
 */
async function handleGET(request: NextRequest, { params }: { params: Promise<{ path: string[]; n: string }> }) {
  const ctx = await createApiContext();
  const raw = await params;

  const parsedPath = parseParams(pathParamSchema, raw.path);
  if (!parsedPath.ok) {
    return apiError('BAD_REQUEST', formatZodError(parsedPath.error), 400);
  }

  const parsedVersion = parseParams(versionParamSchema, raw.n);
  if (!parsedVersion.ok) {
    return apiError('BAD_REQUEST', formatZodError(parsedVersion.error), 400);
  }

  const path = getPathFromParams(raw);

  try {
    const revision = await pageService.getRevision(ctx, path, parsedVersion.data);
    if (!revision) {
      return apiError('NOT_FOUND', 'Revision not found', 404);
    }
    return NextResponse.json(revision);
  } catch (error) {
    if (error instanceof DomainError) return mapDomainError(error);
    return internalError();
  }
}

export const GET = withApiAudit(handleGET as unknown as RouteHandler);