# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Next Wiki API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the Next Wiki REST API."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      }
    },
    "schemas": {
...
      "apiKeyViewListSchema": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/apiKeyViewSchema"
        },
        "description": "List of API keys"
      },
...
    },
...
  },
  "paths": {
...
    "/edit/{path}": {
      "get": {
        "operationId": "get-edit-{path}",
        "summary": "",
        "description": "",
        "tags": [
          "Edit"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: path"
          }
        ],
        "responses": {}
      },
      "post": {
        "operationId": "post-edit-{path}",
        "summary": "",
        "description": "",
        "tags": [
          "Edit"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: path"
          }
        ],
        "responses": {}
      }
    },
...
    "/pages/{path}": {
      "get": {
        "operationId": "get-pages-{path}",
        "summary": "",
        "description": "",
        "tags": [
          "Pages"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: path"
          }
        ],
        "responses": {}
      },
      "delete": {
        "operationId": "delete-pages-{path}",
        "summary": "",
        "description": "",
        "tags": [
          "Pages"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: path"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/properties/{path}": {
      "patch": {
        "operationId": "patch-properties-{path}",
        "summary": "",
        "description": "",
        "tags": [
          "Properties"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: path"
          }
        ],
        "responses": {}
      }
    },
...
    "/revisions/{n}/{path}": {
      "get": {
        "operationId": "get-revisions-{n}-{path}",
        "summary": "",
        "description": "",
        "tags": [
          "Revisions"
        ],
        "parameters": [
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: n"
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: path"
          }
        ],
        "responses": {}
      }
    },
    "/api-keys": {
      "get": {
        "operationId": "get-api-keys",
        "summary": "List API keys",
        "description": "Returns the current user's API keys without secrets.",
        "tags": [
          "User"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiKeyViewListSchema"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-api-keys",
        "summary": "Create an API key",
        "description": "Creates a new API key with the requested scopes. The full secret is returned only once.",
        "tags": [
          "User"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createApiKeyInputSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiKeyCreatedSchema"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{id}/reset-password": {
      "post": {
        "operationId": "post-users-{id}-reset-password",
        "summary": "",
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {}
      }
    },
    "/users/{id}/role": {
      "post": {
        "operationId": "post-users-{id}-role",
        "summary": "",
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {}
      }
    },
    "/users/{id}/status": {
      "post": {
        "operationId": "post-users-{id}-status",
        "summary": "",
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {}
      }
    }
  },
...
}